Move customized widgets from datadialogs.py to widgets.py
